兼容所有Linux4
